renaming the io folder in workflows
