+++ src/rtctools_heat_network/workflows/simulator_workflow.py
@@ -22,7 +22,7 @@ from rtctools.util import run_optimization_problem
 from rtctools_heat_network.esdl.esdl_mixin import ESDLMixin
 from rtctools_heat_network.head_loss_mixin import HeadLossOption
 from rtctools_heat_network.heat_mixin import HeatMixin
-from rtctools_heat_network.workflows.io.write_output import ScenarioOutput
+from rtctools_heat_network.workflows.io_tools.write_output import ScenarioOutput
 from rtctools_heat_network.workflows.utils.adapt_profiles import (
     adapt_hourly_year_profile_to_day_averaged_with_hourly_peak_day,
 )
